// src/frontend/src/backend.d.ts
import type { Principal } from "@icp-sdk/core/principal";
export interface Some<T> {
    __kind__: "Some";
    value: T;
}
export interface None {
    __kind__: "None";
}
export type Option<T> = Some<T> | None;
export interface SearchResult {
    model: string;
    answer: string;
    sources: Array<Source>;
    followUps: Array<string>;
}
export interface TransformationOutput {
    status: bigint;
    body: Uint8Array;
    headers: Array<http_header>;
}
export interface Thread {
    id: string;
    title: string;
    messages: Array<Message>;
    createdAt: bigint;
    results: Array<SearchResult>;
    updatedAt: bigint;
}
export interface TransformationInput {
    context: Uint8Array;
    response: http_request_result;
}
export interface Source {
    url: string;
    title: string;
    snippet: string;
    favicon: string;
}
export interface Message {
    content: string;
    role: string;
    timestamp: bigint;
}
export interface http_header {
    value: string;
    name: string;
}
export interface http_request_result {
    status: bigint;
    body: Uint8Array;
    headers: Array<http_header>;
}
export interface backendInterface {
    createThread(title: string): Promise<string>;
    deleteThread(threadId: string): Promise<boolean>;
    getMyThreads(): Promise<Array<Thread>>;
    getThread(threadId: string): Promise<Thread>;
    listThreads(page: bigint, pageSize: bigint): Promise<Array<Thread>>;
    search(threadId: string, wildcardQuery: string, model: string, focusMode: string): Promise<SearchResult>;
    transform(input: TransformationInput): Promise<TransformationOutput>;
}